// src/shared/types/nutritionchanges.tsx
export type NutritionChanges = {
    calories: number | null
    carbohydrates: number | null
    cholesterol: number | null
    fat: number | null
    fiber: number | null
    protein: number | null
    saturatedFat: number | null
    sodium: number | null
    sugar: number | null
    transFat: number | null
    unsaturatedFat: number | null
}